Render public gallery index statically
